Codec.Binary.Bech32.Internal

Description

Implementation of the Bech32 address format.

From an original implementation by Marko Bencun:

sipa/bech32

Encoding & Decoding

encode :: HumanReadablePart -> DataPart -> Either EncodingError Text

Encode a human-readable string and data payload into a Bech32 string.

encodeLenient :: HumanReadablePart -> DataPart -> Text

Like encode but allows output to be longer than 90 characters. This isn't ideal as the error detection becomes worse as string get longer but it's still acceptable.

From BIP-0173:

Even though the chosen code performs reasonably well up to 1023 characters, other designs are preferable for lengths above 89 characters (excluding the separator).

data EncodingError

Represents the set of error conditions that may occur while encoding a Bech32 string.

Constructors

EncodedStringTooLong

decode :: Text -> Either DecodingError (HumanReadablePart, DataPart)

Decode a Bech32 string into a human-readable part and data part.

decodeLenient :: Text -> Either DecodingError (HumanReadablePart, DataPart)

Like decode but does not enforce a maximum length. See also encodeLenient for details.

data DecodingError

Represents the set of errors that may occur while decoding a Bech32 string with the decode function.

Constructors

StringToDecodeTooLong

StringToDecodeTooShort

StringToDecodeHasMixedCase

StringToDecodeMissingSeparatorChar

StringToDecodeContainsInvalidChars [CharPosition]
In cases where it is possible to determine the exact locations of erroneous characters, this list will encode those locations. Clients can use this information to provide user feedback. In cases where it isn't possible to reliably determine the locations of erroneous characters, this list will be empty.

checksumLength :: Int

The length of the checksum portion of an encoded string, in bytes.

encodedStringMaxLength :: Int

The maximum length of an encoded string, in bytes. This length includes the human-readable part, the separator character, the encoded data portion, and the checksum.

encodedStringMinLength :: Int

The minimum length of an encoded string, in bytes. This length includes the human-readable part, the separator character, the encoded data portion, and the checksum.

separatorChar :: Char

The separator character. This character appears immediately after the human-readable part and before the data part.

separatorLength :: Int

The length of the separator portion of an encoded string, in bytes.

Data Part

data DataPart

Represents the data part of a Bech32 string, as defined here: https://git.io/fj8FS

dataPartIsValid :: DataPart -> Bool

Returns true iff. the specified DataPart is valid.

dataPartFromBytes :: ByteString -> DataPart

Constructs a DataPart from a ByteString.

This function encodes a ByteString in such a way that guarantees it can be successfully decoded with the dataPartToBytes function:

dataPartToBytes (dataPartFromBytes b) == Just b

dataPartFromText :: Text -> Maybe DataPart

Constructs a DataPart from textual input. All characters in the input must be a member of dataCharList, the set of characters permitted to appear within the data part of a Bech32 string.

Returns Nothing if any character in the input is not a member of dataCharList.

This function guarantees to satisfy the following property:

dataPartFromText (dataPartToText d) == Just d

dataPartFromWords :: [Word5] -> DataPart

Construct a DataPart directly from words.

This function guarantees to satisfy the following properties:

dataPartFromWords (dataPartToWords d) == d
dataPartToWords (dataPartFromWords w) == w

dataPartToBytes :: DataPart -> Maybe ByteString

Attempts to extract a ByteString from a DataPart.

This function guarantees to satisfy the following property:

dataPartToBytes (dataPartFromBytes b) == Just b

dataPartToText :: DataPart -> Text

Converts a DataPart to Text, using the Bech32 character set to render the data.

This function guarantees to satisfy the following property:

dataPartFromText (dataPartToText d) == Just d

dataPartToWords :: DataPart -> [Word5]

Convert a DataPart into words.

dataCharToWord :: Char -> Maybe Word5

If the specified character is permitted to appear within the data part of a Bech32 string, this function returns that character's corresponding Word5 value. If the specified character is not permitted, or if the specified character is upper-case, returns Nothing.

dataCharFromWord :: Word5 -> Char

Maps the specified Word5 onto a character that is permitted to appear within the data part of a Bech32 string.

dataCharList :: String

A list of all characters that are permitted to appear within the data part of a Bech32 string. See here for more details: https://git.io/fj8FS

Human-Readable Part

data HumanReadablePart

Represents the human-readable part of a Bech32 string, as defined here: https://git.io/fj8FS

data HumanReadablePartError

Represents the set of error conditions that may occur while parsing the human-readable part of a Bech32 string.

Constructors

HumanReadablePartTooShort

HumanReadablePartTooLong

HumanReadablePartContainsInvalidChars [CharPosition]

humanReadablePartFromText :: Text -> Either HumanReadablePartError HumanReadablePart

Parses the human-readable part of a Bech32 string, as defined here: https://git.io/fj8FS

humanReadablePartToText :: HumanReadablePart -> Text

Get the raw text of the human-readable part of a Bech32 string.

humanReadablePartToWords :: HumanReadablePart -> [Word5]

Convert the specified human-readable part to a list of words.

humanReadablePartMinLength :: Int

The shortest length permitted for the human-readable part of a Bech32 string.

humanReadablePartMaxLength :: Int

The longest length permitted for the human-readable part of a Bech32 string.

humanReadableCharMinBound :: Char

The lower bound of the set of characters permitted to appear within the human-readable part of a Bech32 string.

humanReadableCharMaxBound :: Char

The upper bound of the set of characters permitted to appear within the human-readable part of a Bech32 string.

Bit Manipulation

convertBits :: Functor f => [Word] -> Int -> Int -> Pad f -> f [Word]

Big-endian conversion of a word string from base '2^frombits' to base '2^tobits'. The frombits and twobits parameters must be positive, while '2^frombits' and '2^tobits' must be smaller than the size of Word. Every value in dat must be strictly smaller than '2^frombits'.

data Word5

word5 :: Integral a => a -> Word5

getWord5 :: Word5 -> Word8

toBase256 :: [Word5] -> Maybe [Word8]

toBase32 :: [Word8] -> [Word5]

noPadding :: Pad Maybe

yesPadding :: Pad Identity

Character Manipulation

newtype CharPosition

The zero-based position of a character in a string, counting from the left.

Constructors

CharPosition Int