Codec.Binary.Bech32

Description

Implementation of the Bech32 address format.

From an original implementation by Marko Bencun:

sipa/bech32

Encoding & Decoding

encode :: HumanReadablePart -> DataPart -> Either EncodingError Text

Encode a human-readable string and data payload into a Bech32 string.

decode :: Text -> Either DecodingError (HumanReadablePart, DataPart)

Decode a Bech32 string into a human-readable part and data part.

Data Part

data DataPart

Represents the data part of a Bech32 string, as defined here: https://git.io/fj8FS

dataPartFromBytes :: ByteString -> DataPart

Constructs a DataPart from a ByteString.

This function encodes a ByteString in such a way that guarantees it can be successfully decoded with the dataPartToBytes function:

dataPartToBytes (dataPartFromBytes b) == Just b

dataPartFromText :: Text -> Maybe DataPart

Constructs a DataPart from textual input. All characters in the input must be a member of dataCharList, the set of characters permitted to appear within the data part of a Bech32 string.

Returns Nothing if any character in the input is not a member of dataCharList.

This function guarantees to satisfy the following property:

dataPartFromText (dataPartToText d) == Just d

dataPartToBytes :: DataPart -> Maybe ByteString

Attempts to extract a ByteString from a DataPart.

This function guarantees to satisfy the following property:

dataPartToBytes (dataPartFromBytes b) == Just b

dataPartToText :: DataPart -> Text

Converts a DataPart to Text, using the Bech32 character set to render the data.

This function guarantees to satisfy the following property:

dataPartFromText (dataPartToText d) == Just d

Human-Readable Part

data HumanReadablePart

Represents the human-readable part of a Bech32 string, as defined here: https://git.io/fj8FS

humanReadablePartFromText :: Text -> Either HumanReadablePartError HumanReadablePart

Parses the human-readable part of a Bech32 string, as defined here: https://git.io/fj8FS

humanReadablePartToText :: HumanReadablePart -> Text

Get the raw text of the human-readable part of a Bech32 string.