Codec.Binary.Bech32.Internal

Description

Implementation of the Bech32 address format.

From an original implementation by Marko Bencun:

sipa/bech32

Encoding & Decoding

encode :: HumanReadablePart -> ByteString -> Maybe ByteString

decode :: ByteString -> Maybe (HumanReadablePart, ByteString)

checksumLength :: Int

The length of the checksum portion of an encoded string, in bytes.

maxEncodedStringLength :: Int

The maximum length of an encoded string, in bytes. This length includes the human-readable part, the separator character, the encoded data portion, and the checksum.

Human-Readable Parts

data HumanReadablePart

mkHumanReadablePart :: ByteString -> Maybe HumanReadablePart

humanReadablePartToBytes :: HumanReadablePart -> ByteString

humanReadableCharsetMinBound :: Word8

humanReadableCharsetMaxBound :: Word8

Bit Manipulation

convertBits :: Functor f => [Word] -> Int -> Int -> Pad f -> f [Word]

data Word5

word5 :: Integral a => a -> Word5

getWord5 :: Word5 -> Word8

toBase256 :: [Word5] -> Maybe [Word8]

toBase32 :: [Word8] -> [Word5]

noPadding :: Pad Maybe

yesPadding :: Pad Identity

Character Set Manipulation

charset :: Array Word5 Char

charsetMap :: Char -> Maybe Word5